[iOS] feat: Setting 과 User Network Manager 분리
