+++ iOS/DaDaIkSeon/DaDaIkSeon/Network/User/UserEndpoint.swift
@@ -11,29 +11,20 @@ enum UserEndpoint {
     case get
     case postEmail(email: String)
     case postCode(code: String, email: String, device: Device)
-    case patchEmail(email: String)
-    case patchBackup(isBackup: Bool, udid: String)
-    case patchMultiDevice(isMultiDevice: Bool)
-//    case
-//
 }
 
 extension UserEndpoint: EndpointType {
     
     var path: String {
-        let basePath = baseUrl
+        let basePath = baseUrl + "/app/user"
         
         switch self {
         case .get:
             return basePath
-        case .postEmail, .patchEmail:
+        case .postEmail:
             return basePath + "/email"
         case .postCode:
             return basePath + "/confirm-email"
-        case .patchBackup(let id, _):
-            return basePath + "/backup/\(id)"
-        case .patchMultiDevice:
-            return basePath + "/multi"
         }
     }
     
@@ -43,8 +34,6 @@ extension UserEndpoint: EndpointType {
             return .GET
         case .postEmail, .postCode:
             return .POST
-        case .patchEmail, .patchBackup, .patchMultiDevice:
-            return .PATCH
         }
     }
     
@@ -52,7 +41,7 @@ extension UserEndpoint: EndpointType {
         switch self {
         case .get:
             return nil
-        case .postEmail(let email), .patchEmail(let email):
+        case .postEmail(let email):
             return ["email": email]
         case .postCode(let code, let email, let device):
             return [
@@ -60,10 +49,6 @@ extension UserEndpoint: EndpointType {
                 "email": email,
                 "device": device
             ]
-        case .patchBackup(_, let backup):
-            return ["backup": backup]
-        case .patchMultiDevice(let multiDevice):
-            return ["multiDevice": multiDevice]
         }
     }
