+++ iOS/DaDaIkSeon/DaDaIkSeon/Network/User/UserNetworkManager.swift
@@ -44,53 +44,4 @@ final class UserNetworkManager: Requestable {
         }
     }
     
-    func changeEmail(email: String,
-                     completion: @escaping () -> Void) {
-        
-        userEndpoint = .patchEmail(email: email)
-        request(userEndpoint) { result in
-            switch result {
-            case .networkSuccess:
-                completion()
-            case .networkError(let error):
-                print(error)
-            case .networkFail:
-                print("Network Fail!!!!")
-            }
-        }
-    }
-    
-    func changeBackupMode(backup: Bool,
-                          udid: String,
-                          completion: @escaping () -> Void) {
-        
-        userEndpoint = .patchBackup(isBackup: backup,
-                                    udid: udid)
-        request(userEndpoint) { result in
-            switch result {
-            case .networkSuccess:
-                completion()
-            case .networkError(let error):
-                print(error)
-            case .networkFail:
-                print("Network Fail!!!!")
-            }
-        }
-    }
-
-    func changeMultiDevice(multiDevice: Bool,
-                           completion: @escaping () -> Void) {
-        
-        userEndpoint = .patchMultiDevice(isMultiDevice: multiDevice)
-        request(userEndpoint) { result in
-            switch result {
-            case .networkSuccess:
-                completion()
-            case .networkError(let error):
-                print(error)
-            case .networkFail:
-                print("Network Fail!!!!")
-            }
-        }
-    }
 }
